Refactoring page structure
