+++ activity_browser/layouts/pages/activity_details/models.py
@@ -1,0 +1,7 @@
+from activity_browser.ui.widgets import ABAbstractItemModel
+
+from .items import ExchangeItem
+
+
+class ExchangeModel(ABAbstractItemModel):
+    dataItemClass = ExchangeItem
